% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_groups.R
\name{age_groups}
\alias{age_groups}
\title{Age groups}
\usage{
age_groups(age)
}
\arguments{
\item{age}{The students' ages}
}
\value{
a factor of age groups
}
\description{
Age groups
}
\details{
The recoded age groups are Under 21, 21-25, 26-30, 31-35, 36-40, and 41 and above.
}
