% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_credit.R
\name{dual_credit}
\alias{dual_credit}
\title{Identify students in dual-credit programs}
\usage{
dual_credit(Id, Program, term = NULL, term_num = NULL)
}
\arguments{
\item{Id}{Student ID}

\item{Program}{Academic program code, usually from ST Students Gender/Ethnicity by Program}

\item{term}{The academic term (e.g., 2020SP). This argument is optional if \code{term_num} is supplied.}

\item{term_num}{The numeric code corresponding to the academic term (e.g., 49 for 2020SP). This argument is optional if \code{term} is supplied.}
}
\value{
A data.frame containing \code{Id}, \code{Program},  \code{term_num},
\code{term} (if supplied),  a logical flag indicating if \code{Program} is a dual-credit program (\code{dual_credit}), a logical flag indicating if any of the programs the student took in each term where dual-credit programs (\code{any_dual_credit}), and a logical flag indicating if all of the programs the student took in each term where dual-credit programs (\code{all_dual_credit}).
}
\description{
Identify students in dual-credit programs
}
\details{
The method to determine if a program is dual-credit is explained in this table:
\tabular{lll}{
Term              \tab Program Code(s)   \tab Dual-credit?\cr
2015FL and before \tab EHDC & EHSE       \tab Yes\cr
2015FL and before \tab EHSE              \tab Yes\cr
2015FL and before \tab EHDC              \tab Yes\cr
2015FL and before \tab \emph{ECHS} & EHDC     \tab No \cr
2015FL and before \tab \emph{ECHS} & EHSE     \tab No \cr
2015FL and before \tab \emph{ECHS}            \tab No \cr
2015FL and before \tab None of the above \tab No \cr
After 2015FL      \tab EHSE              \tab Yes\cr
After 2015FL      \tab EHDC & EHSE       \tab Yes\cr
After 2015FL      \tab \emph{ECHS} & EHSE     \tab No \cr
After 2015FL      \tab \emph{ECHS}            \tab No \cr
After 2015FL      \tab \emph{ECHS} & EHDC     \tab No \cr
After 2015FL      \tab EHDC              \tab No \cr
After 2015FL      \tab None of the above \tab No
}
}
