% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race_and_ethnicity_recode.R
\name{ethnic_recode}
\alias{ethnic_recode}
\title{ethnic_recode}
\usage{
ethnic_recode(Races, Ethnics)
}
\arguments{
\item{Ethnics}{}
}
\value{
a character vector.
}
\description{
Calculates the new race/ethnicity definition (07-19-2017).  Based on \verb{New Ethnicity Recodes for ST Students Enrollment Files 10-25-2018.sps}. Updated 2020-09-15 to account for Ethnics = "No Ethnicity Listed".
}
