% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Term_to_numeric_value.R
\name{factor_term}
\alias{factor_term}
\title{Get factor-ized terms}
\usage{
factor_term(Term, drop = c("outside", "all", "none"), Qs = TRUE)
}
\arguments{
\item{Term}{A variable containing academic terms.}

\item{drop}{Character. How should term levels not seen in the data be handled? Default is \code{drop = "outside"}, which removes term levels outside of the range seen in the data but keeps unobserved terms occurring between the first and last observed terms. Using \code{drop = "all"} removes all term levels that are not present in the data (may create 'gaps' in the term levels).  Using \code{drop = "none"} will keep all term levels between 2004SP and 2050Q2.}

\item{Qs}{Logical.  Should quarters be included in the term levels? Overrides the \code{drop} argument.}
}
\value{
A factor with the term levels ordered to allow sorting.
}
\description{
\code{factor_term()} takes a Term column/variable and returns it as a factor for sorting purposes.
}
