% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_generation.R
\name{first_gen}
\alias{first_gen}
\alias{first_generation}
\alias{first_generation_orbit}
\alias{first_generation_discoverer}
\title{First Generation Student}
\usage{
first_generation(
  Mother_Educ_Level,
  Father_Educ_Level,
  method = c("tcc", "fed"),
  parents = FALSE,
  source = c("Discoverer", "Orbit")
)

first_generation_orbit(
  Mother_Educ_Level,
  Father_Educ_Level,
  method = c("tcc", "fed"),
  parents = FALSE,
  warn = TRUE
)

first_generation_discoverer(
  Mother_Educ_Level,
  Father_Educ_Level,
  method = c("tcc", "fed"),
  parents = FALSE,
  warn = TRUE
)
}
\arguments{
\item{Mother_Educ_Level}{The highest level of educational attainment of the
student's mother.}

\item{Father_Educ_Level}{The highest level of educational attainment of the
student's father.}

\item{method}{What definition of "first generation student" to use. Use
\code{method = "tcc"} to use the internal definition (student is first
generation if mom \emph{and} dad have no college); this definition is
mainly used internally for reports such as the extended campus profiles.
Use \code{method = "fed"} to use the federal definition (student is first
generation if mom \emph{and} dad have no bachelor's degree); this
definition is used for grant requests.}

\item{parents}{Logical. Should the re-coded values for parents' educational
levels be returned? Default is FALSE.}

\item{source}{Either a character of length one indicating the data source for
the entire dataset ("Discoverer" (the default) or "Orbit"), or a character
vector indicating the data source for each observation.
\strong{Note}: For 2020FL and later, use data from Orbit.}
}
\value{
if \code{parents = FALSE} (the default), a vector indicating if
students are first generation. If \code{parents = TRUE}, a dataframe
containing mothers' re-coded educational level, fathers' re-coded
educational level, and students' first generation status.
}
\description{
First Generation Student
}
\details{
The value of the argument \code{source} will determine which of two auxiliary functions gets called (\code{first_generation_discoverer} or \code{first_generation_orbit}).

\strong{Definitions}
\itemize{
\item NOHS = No High School
\item SMHS = Some High School
\item HSGD = HS Grad or GED
\item SMCL = Some College
\item CLBD = Completed Bachelor's Degree
\item CLAD = Completed Advanced Degree
\item NRSP = No Response
\item LESS = Less than High School
\item GED = GED
\item HIGH = High School
\item VOC = Vocational/Technical
\item SOME = Some College
\item ASSOC = Associates Degree
\item BACH = Bachelors Degree
\item GRAD = Some Graduate School
\item MAST = Masters or Professional
\item DOC = Doctoral Degree
\item UNKNOWN = Unknown
}
}
\note{
For 2020FL and later, use data from Orbit.
}
