% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_generation.R
\name{first_generation}
\alias{first_generation}
\title{First Generation Student}
\usage{
first_generation(
  Mother_Educ_Level,
  Father_Educ_Level,
  method = "tcc",
  parents = FALSE
)
}
\arguments{
\item{Mother_Educ_Level}{The highest level of educational attainment of the
student's mother.}

\item{Father_Educ_Level}{The highest level of educational attainment of the
student's father.}

\item{method}{What definition of "first generation student" to use. Use
\code{method = "tcc"} to use the internal definition (student is first
generation if mom \emph{and} dad have no college); this definition is
mainly used internally for reports such as the extended campus profiles.
Use \code{method = "fed"} to use the federal definition (student is first
generation if mom \emph{and} dad have no bachelor's degree); this
definition is used for grant requests.}

\item{parents}{Logical. Should the re-coded values for parents' educational
levels be returned? Default is FALSE.}
}
\value{
if \code{parents = FALSE} (the default), a vector indicating if
students are first generation. If \code{parents = TRUE}, a dataframe
containing mothers' re-coded educational level, fathers' re-coded
educational level, and students' first generation status.
}
\description{
First Generation Student
}
\details{
Defintions:
NOHS = No High School
SMHS = Some High School
HSGD = HS Grad or GED
SMCL = Some College
CLBD = Completed Bachelor's Degree
CLAD = Completed Advanced Degree
NRSP = No Response
}
