% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd.R
\name{isd}
\alias{isd}
\title{ISD}
\usage{
isd(High_School)
}
\arguments{
\item{High_School}{Character. A the high schools to get ISDs for.}
}
\value{
a \code{data.frame} with the variables \code{ISD}, \code{Include_ISD}, \code{Include_ALL}, \code{Service_Area}, and \code{Service_Area_with_Private_Charter}
}
\description{
ISD
}
