% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{like}
\alias{like}
\alias{\%like\%}
\alias{\%LIKE\%}
\title{Like}
\usage{
x \%like\% y

x \%LIKE\% y
}
\arguments{
\item{x}{a vector that can be coerced to character: the values to be matched against.}

\item{y}{a scalar that can be coerced to character: the value to be matched.}
}
\description{
Like
}
\details{
The functions \code{\%like\%} and \code{\%LIKE\%} are
identical; the later is supplied so I don't have to lift my finger off
the shift key. Uses \link[stringr]{str_detect} from \code{stringr} to allow
use of \code{stringr}'s \link[stringr:modifiers]{stringr::modifiers}, such as \code{fixed}.

The matching function ignores the case of the characters.
}
\examples{
LETTERS[1:5] \%like\% "A"  #TRUE FALSE FALSE FALSE FALSE
LETTERS[1:5] \%LIKE\% "A"  #Same
LETTERS[1:5] \%LIKE\% stringr::fixed("a", ignore_case = T)  #Same

vec = c("")


}
