% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_print.R
\name{my_print}
\alias{my_print}
\title{Print data in a pretty and helpful way}
\usage{
my_print(df, title = NULL, ...)
}
\arguments{
\item{df}{a data frame (or similar) that you want to print}

\item{title}{a (character) string to use for the table}

\item{...}{other arguments}
}
\description{
Print data in a pretty and helpful way
}
\seealso{
\code{?DT::datatable}
}
