% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race_and_ethnicity_recode.R
\name{new_ethnic_group}
\alias{new_ethnic_group}
\title{new_ethnic_group}
\usage{
new_ethnic_group(newEthnicity)
}
\arguments{
\item{newEthnicity}{the result from running \code{ethnic_recode()}

#' @description Calculates the new race/ethnicity definition (07-19-2017).  Based on \verb{New Ethnicity Recodes for ST Students Enrollment Files 10-25-2018.sps}.}
}
\value{
a character vector.
}
\description{
new_ethnic_group
}
\examples{
new_ethnic_group("White") #"1. White"
new_ethnic_group("Black / African American") #"2. Black / African American"
new_ethnic_group("Hispanic Latino") #"3. Hispanic Latino"
new_ethnic_group("Asian") #"4. Other"   everything else
}
