% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race_and_ethnicity_recode.R
\name{newEthnicity3}
\alias{newEthnicity3}
\title{newEthnicity3}
\usage{
newEthnicity3(Races, Ethnics)
}
\arguments{
\item{Races}{The 'Races' variable.}

\item{Ethnics}{The 'Ethnics' variable.}
}
\description{
Calculates the new race/ethnicity definition (07-19-2017).  Based on \verb{New Ethnicity Recodes for ST Students Enrollment Files 10-25-2018.sps}.
}
