% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nin}
\alias{nin}
\alias{\%nin\%}
\alias{\%NIN\%}
\title{Not in}
\usage{
x \%nin\% table

x \%NIN\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are
supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors
are not supported.}
}
\description{
Not in
}
\details{
Similar to the function \code{\%in\%} from \code{base} but returns true if
there is \strong{no} match. The functions \code{\%nin\%} and \code{\%NIN\%} are
identical; the later is supplied so I don't have to lift my finger off
the shift key. Uses \link[fastmatch]{fmatch} from \code{fastmatch} for speed.
}
\seealso{
\link[base]{match} from \code{base} and \link[fastmatch]{fmatch} from \code{fastmatch}
}
