% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_list}
\alias{print_list}
\title{Print a list of values to console}
\usage{
print_list(x, method = c("discoverer", "orbit"))
}
\arguments{
\item{x}{The values to be turned into a list}

\item{method}{Where is the list to be used? If \code{method = "discoverer"}
(default for now), the list elements are surrounded by single quotes and
separated by commas.  If \code{method = "orbit"}, the list elements are
surrounded by double quotes and separated by semicolons.}
}
\description{
Print a list of values to console
}
\examples{
print_list(1:9, method = "disc") # '1', '2', '3', '4', '5', '6', '7', '8', '9'

print_list(1:9, method = "o") # "1"; "2"; "3"; "4"; "5"; "6"; "7"; "8"; "9"
}
