% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_read.R
\name{read_folder}
\alias{read_folder}
\title{Read in a folder of data}
\usage{
read_folder(path, clean_names = T, blanks_to_na = T, ...)
}
\arguments{
\item{path}{A character string containing a single file name for a multi-object file (e.g., folder/directory containing data files, Excel workbook with multiple sheets, zip directory, or HTML file), or a vector of file paths for multiple files to be imported.}

\item{clean_names}{Logical. Should the variable names be cleaned? Default is TRUE.}

\item{blanks_to_na}{Logical. Should blank/empty character strings be converted to NAs? Default is TRUE.}

\item{...}{Other arguments passed to methods}
}
\value{

}
\description{
Read in a folder of data
}
\examples{

}
