% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCC_color_scales.R
\name{scale_color_tccd}
\alias{scale_color_tccd}
\title{Color scale constructor for TCCD colors}
\usage{
scale_color_tccd(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tccd_palettes}

\item{discrete}{Boolean indicating if color aesthetic is discrete}

\item{reverse}{Boolean indicating if the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for TCCD colors
}
