% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_ids.R
\name{standardize_ids}
\alias{standardize_ids}
\title{Standardize the length of ID strings}
\usage{
standardize_ids(id, length = 7)
}
\arguments{
\item{id}{a numeric vector (or one that can be coerced to numeric) containing IDs}

\item{length}{a scalar integer indicating how long the result should be.  Default is 7.}
}
\value{
a character vector of IDs with leading zeros added
}
\description{
Standardizes the length of IDs (or other variables) by adding leading zeros.
}
