% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCCD.R
\docType{package}
\name{TCCD-package}
\alias{TCCD}
\alias{TCCD-package}
\title{TCCD: R Functions for Work at TCCD}
\description{
A collection of R functions for work done at TCCD.
}
\author{
\strong{Maintainer}: Jacob Roden-Foreman \email{jakerodenforeman@gmail.com} (\href{https://orcid.org/0000-0002-8338-397X}{ORCID})

}
\keyword{internal}
