% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCC_color_scales.R
\name{tccd_cols}
\alias{tccd_cols}
\title{Function to extract TCCD colors as hex codes}
\usage{
tccd_cols(...)
}
\arguments{
\item{...}{Character names of tccd_colors}
}
\description{
Function to extract TCCD colors as hex codes
}
