% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCC_color_scales.R
\name{tccd_pal}
\alias{tccd_pal}
\title{Return function to interpolate a TCCD color palette}
\usage{
tccd_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tccd_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to colorRampPalette()}
}
\description{
Return function to interpolate a TCCD color palette
}
