% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_to_acad_year.R
\name{term_ID_to_acad_year}
\alias{term_ID_to_acad_year}
\title{term_ID_to_acad_year}
\usage{
term_ID_to_acad_year(Term_Id)
}
\arguments{
\item{Term_Id}{A character vector that identifies the term (e.g., "2019FL" or "2019Q1").}
}
\value{
A character vector of the academic year that the terms belong to (e.g., ""2019-2020).
}
\description{
Based on \code{AcademicYear.sps}
}
\examples{
term_ID_to_acad_year("2019SU")   # "2018-2019"

}
