% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Term_to_numeric_value.R
\name{term_num}
\alias{term_num}
\title{Term_num}
\usage{
term_num(TERM_ID)
}
\arguments{
\item{TERM_ID}{A term character or factor code with values such as "2019FL" or "2010SU".}
}
\description{
Converts terms to a numeric value for calculations.  Based on \verb{Convert STCTERM to numeric value.sps}.
}
\details{
Convert Terms to a numeric value for calculations.
}
