% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_to_acad_year.R
\name{term_num_to_acad_year}
\alias{term_num_to_acad_year}
\title{term_num_to_acad_year}
\usage{
term_num_to_acad_year(Term_num)
}
\arguments{
\item{Term_num}{the numeric code for the term, such as calculated by \code{Term_num()}}
}
\value{
A character vector of the relevant academic year (e.g., "2018-2019").
}
\description{
term_num_to_acad_year
}
