% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCC_color_scales.R
\name{theme_tcc}
\alias{theme_tcc}
\title{TCC Theme For GGPlot2}
\usage{
theme_tcc(
  rotate_x_labs = FALSE,
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{rotate_x_labs}{Logical. Should the labels for the x axis be rotated 90 degrees? Default is FALSE.}

\item{base_size}{Numeric. Base font size. Default is 11.}

\item{base_family}{Character. Base font family.}

\item{base_line_size}{Numeric. Base size for line elements.}

\item{base_rect_size}{Numeric. Base size for rect elements.}
}
\description{
A TCC theme for ggplot2.  Based on \link[ggplot2]{theme_bw}
}
\examples{

p <- ggplot(mpg) +
	geom_bar(aes(y = manufacturer, fill = drv)) +
	scale_fill_tccd("blue_green") +
	#No space below the bars but 10\% above them
	scale_x_continuous(expand = expansion(mult = c(0, 0.1))) +
	coord_flip() +
	labs(
		title  = "This is  a chart",
		subtitle = "It shows you some data",
		caption = "This is a caption. Some important notes could go here."
	)

p

p + theme_bw()

#based on theme_bw() but aligns the caption with the plot instead of the figure
# and it adds some extra space between the x axis title and the caption
p + theme_tcc()

#easily rotate and align x axis labels
p + theme_tcc(rotate_x_labs = T)
}
\seealso{
\link[ggplot2]{theme_bw}
}
