% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%IN\%}
\alias{\%IN\%}
\title{In}
\usage{
x \%IN\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are
supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors
are not supported.}
}
\description{
In
}
\details{
This is the exact same as \code{\%fin\%} from \code{fastmatch},
but this way I don't have to lift my finger off the shift key.
}
\seealso{
\link[fastmatch]{fmatch} from \code{fastmatch}
}
