% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{like}
\alias{like}
\alias{\%like\%}
\alias{\%LIKE\%}
\title{Like}
\usage{
x \%like\% y

x \%LIKE\% y
}
\arguments{
\item{x}{a vector that can be coerced to character: the values to be matched against.}

\item{y}{a scalar that can be coerced to character: the value to be matched.}
}
\description{
Like
}
\details{
The functions \code{\%like\%} and \code{\%LIKE\%} are
identical; the later is supplied so I don't have to lift my finger off
the shift key.

The matching function ignores the case of the characters.
}
\examples{
LETTERS[1:5] \%like\% "A"  #TRUE FALSE FALSE FALSE FALSE
LETTERS[1:5] \%LIKE\% "A"  #TRUE FALSE FALSE FALSE FALSE
LETTERS[1:5] \%like\% "a"  #TRUE FALSE FALSE FALSE FALSE
LETTERS[1:5] \%like\% "C"  #FALSE FALSE TRUE FALSE FALSE
LETTERS[1:5] \%like\% "X"  #FALSE FALSE FALSE FALSE FALSE
"abcdX" \%like\% "X"       #TRUE

}
