% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veteran_status.R
\name{vet_status}
\alias{vet_status}
\title{Veteran and Military-Connected Status}
\usage{
vet_status(Military_Statuses, Smv_Vetben_Elig_Code)
}
\arguments{
\item{Military_Statuses}{Military status code, usually from Student Summary General}

\item{Smv_Vetben_Elig_Code}{Veteran benefits eligibility code, usually from Student Summary General}
}
\value{
Returns a character vector indicating if student is a veteran, dependent, or not-military connected
}
\description{
Veteran and Military-Connected Status
}
